Update LoginTests
Remove duplicate framework, consolidate on the better implementation and better name.
