--- source/test/TestApp/PlayFabPlatformSpecificTest_default.cpp
@@ -1,40 +1,0 @@
-// Copyright (C) Microsoft Corporation. All rights reserved.
-
-// This file is meant to be compiled into projects that do not have a useful/valid platform specific login.
-// When present, add instead the platform specific file, and do NOT add this one
-
-#include "TestAppPch.h"
-
-#include "TestContext.h"
-#include "PlayFabPlatformSpecificTest.h"
-
-namespace PlayFabUnit
-{
-    void PlayFabPlatformSpecificTest::TestPlatformSpecificLogin(TestContext&)
-    {
-    }
-
-    void PlayFabPlatformSpecificTest::AddTests()
-    {
-    }
-
-    void PlayFabPlatformSpecificTest::ClassSetUp()
-    {
-    }
-
-    void PlayFabPlatformSpecificTest::SetUp(TestContext&)
-    {
-    }
-
-    void PlayFabPlatformSpecificTest::Tick(TestContext&)
-    {
-    }
-
-    void PlayFabPlatformSpecificTest::TearDown(TestContext&)
-    {
-    }
-
-    void PlayFabPlatformSpecificTest::ClassTearDown()
-    {
-    }
-}
